requirements.txt
streamlit
opencv-python
pytesseract
numpy
Pillow
qrcode
decode
easyocr